+++ apps/sim/app/workspace/[workspaceId]/knowledge/components/constants.ts
@@ -5,16 +5,3 @@ export const dropdownContentClass =
   'w-[220px] rounded-lg border-[var(--surface-6)] bg-[var(--white)] p-0 shadow-xs dark:border-[var(--border-muted)] dark:bg-[var(--surface-2)]'
 
 export const commandListClass = 'overflow-y-auto overflow-x-hidden'
-
-export type SortOption = 'name' | 'createdAt' | 'updatedAt' | 'docCount'
-export type SortOrder = 'asc' | 'desc'
-
-export const SORT_OPTIONS = [
-  { value: 'updatedAt-desc', label: 'Last Updated' },
-  { value: 'createdAt-desc', label: 'Newest First' },
-  { value: 'createdAt-asc', label: 'Oldest First' },
-  { value: 'name-asc', label: 'Name (A-Z)' },
-  { value: 'name-desc', label: 'Name (Z-A)' },
-  { value: 'docCount-desc', label: 'Most Documents' },
-  { value: 'docCount-asc', label: 'Least Documents' },
-] as const
